requirements.txt
streamlit
streamlit_option_menu
langchain-community
langchain-openai
langchain-groq
langchain-core
langchain
python-dotenv
pandas
plotly
numpy
regex
streamlit_chat
google-auth
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
st-gsheets-connection
